requirements.txt
boto3==1.25.1
opencv_python==4.6.0.66
pandas==1.4.2
python-dotenv==0.21.0
